requirements.txt
Flask==3.0.3
gunicorn==21.2.0
google-cloud-firestore==2.16.0
google-cloud-tasks==2.16.3
google-auth==2.34.0
google-api-python-client==2.143.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
